% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atanpow.R
\name{qatanpow}
\alias{qatanpow}
\alias{fatanpow}
\alias{dqatanpow}
\alias{patanpow}
\alias{datanpow}
\title{Arctan Power distribution}
\usage{
qatanpow(p, alpha = 1, beta = 1)

fatanpow(p, alpha = 1, beta = 1)

dqatanpow(p, alpha = 1, beta = 1, log = FALSE)

patanpow(x, alpha = 1, beta = 1)

datanpow(x, alpha = 1, beta = 1)
}
\arguments{
\item{p}{vector of probabilities}

\item{alpha}{shape parameter alpha of Arctan Power Distribution.}

\item{beta}{shape parameter beta of Arctan Power Distribution.}

\item{log}{logical; if TRUE, log density is returned. Default is FALSE}

\item{x}{data}
}
\value{
vector
}
\description{
Defines quantile function (ICDF), and quantile density function as well as probability function (CDF) for Arctan Power distribution
}
\examples{
qatanpow(0.1, 2,2)
}
\references{
Nasiru S, Abubakari AG, Chesneau C. The Arctan Power Distribution: Properties, Quantile and Modal Regressions with Applications to Biomedical Data. Mathematical and Computational Applications. 2023; 28(1):25. https://doi.org/10.3390/mca28010025
}
