% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebyshev.R
\name{get_chebyshev_coeffs}
\alias{get_chebyshev_coeffs}
\alias{find_chebyshev_roots}
\alias{eval_chebyshev_poly}
\alias{fasteval_chebyshev_poly}
\title{Approximate a function by Chebyshev polynomials and find proxy roots}
\usage{
get_chebyshev_coeffs(fun, n, from = 0, to = 1, ..., adjusted = FALSE)

find_chebyshev_roots(a, from = 0, to = 1)

eval_chebyshev_poly(a, x, from = 0, to = 1)

fasteval_chebyshev_poly(a, x, from = 0, to = 1)
}
\arguments{
\item{fun}{function to evaluate}

\item{n}{degree of Chebyshev polynomial}

\item{from, to}{the interval for the input argument}

\item{...}{other arguments passed to a function}

\item{adjusted}{should Chebyshev polynomial be adjusted}

\item{a}{numeric vector of Chebyshev coefficients}

\item{x}{numeric vector of values X to evaluate approximating polynomials for}
}
\value{
a vector of Chebyshev polynomial coefficients
}
\description{
Approximate a function by Chebyshev polynomials and find proxy roots
}
\examples{
get_chebyshev_coeffs(qpd::fexp, lambda=0.25, n=5)
}
