% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chen.R
\name{chen}
\alias{chen}
\alias{qchen}
\alias{fchen}
\alias{dqchen}
\alias{pchen}
\alias{rchen}
\title{Chen distribption}
\usage{
qchen(p, lambda, beta)

fchen(x, lambda, beta, log = FALSE)

dqchen(p, lambda, beta, log = FALSE)

pchen(x, lambda, beta)

fchen(x, lambda, beta, log = FALSE)

rchen(n, lambda, beta)
}
\arguments{
\item{p}{vector of probabilities}

\item{lambda}{shape parameter of Chen distribution, must be positive}

\item{beta}{shape parameter of Chen distribution, must be positive}

\item{x}{numeric observations}

\item{log}{logical; if TRUE, log density is returnes. Default is FALSE}

\item{n}{numeric; number of samples to draw from Tukey Lambda distribution}
}
\value{
vector
}
\description{
Defines quantile function (ICDF), and quantile density function as well as probability function (CDF) for Chen distribution
Distribution range is \eqn{0<x<\infty}
Chen distribution QF, QDF, CDF and PDF are
\deqn{Q(u)=\left[\ln\left(1-\frac{\ln(1-u)}{\lambda}\right)\right]^{1/\beta}}
\deqn{q(u)=\frac{\ln\left(1-\frac{\ln(1-u)}{\lambda}\right)^{\frac{1}{\beta}-1}}{\beta(\ln(1-u)-\lambda)(u-1)}}
\deqn{F(x)=1-\exp(\lambda(1-\exp(x^\beta)))}
\deqn{f(x)=\lambda\beta x^{\beta-1}\exp[\lambda(1-\exp(x^\beta)+x^\beta)]}
}
\examples{
qchen(0.1, 0.5, 1)
p <- runif(1e4)
x <- qchen(p, 0.5, 1)+qchen(p, 1, 0.5)
hist(x,30)
}
\references{
Chen Z (2000) A new two-parameter lifetime distribution with bathtub shape or increasing failure rate function, Statistics & Probability Letters, Volume 49, Issue 2, https://doi.org/10.1016/S0167-7152(00)00044-4.
}
