% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coth}
\alias{coth}
\title{internal function for hyperbolic cotangent borrowed from \code{pracma}}
\usage{
coth(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
vector of hyperbolic secants
}
\description{
internal function for hyperbolic cotangent borrowed from \code{pracma}
}
\keyword{internal}
