% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csch}
\alias{csch}
\title{internal function for hyperbolic cosecant borrowed from \code{pracma}}
\usage{
csch(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
vector of hyperbolic secants
}
\description{
internal function for hyperbolic cosecant borrowed from \code{pracma}
}
\keyword{internal}
