% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diago}
\alias{diago}
\title{Diagonal matrix with offset}
\usage{
diago(x = 1, nrow = NA, ncol = NA, names = TRUE, offset = 1L)
}
\arguments{
\item{x}{a matrix, vector or 1D array, or missing.}

\item{nrow, ncol}{optional dimensions for the result when x is not a matrix.}

\item{names}{(when x is a matrix) logical indicating if the resulting vector,
the diagonal of x, should inherit names from dimnames(x) if available.}

\item{offset}{an integer value to offset the diagonal by. Default is 1}
}
\value{
a square matrix
}
\description{
Diagonal matrix with offset
}
