% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{rdir}
\alias{rdir}
\alias{ddir}
\title{Dirichlet distribution}
\usage{
rdir(n, k)

ddir(x, k)
}
\arguments{
\item{n}{number of observations to draw from the Dirichlet distribution}

\item{k}{vector of parameters}

\item{x}{matrix \code{n*m} of \code{n} (rows) simplices of dimension \code{m} (columns)}
}
\value{
returns a matrix of samples from the Dirichlet distribution
with number of rows n for \code{rdir()} or a vector or densities for \code{ddir()}
}
\description{
Dirichlet distribution

Title
}
\examples{
k <- c(3.14, 10.6, 2.24, 8.67)
rdir(20, k)
x <- rdir(5, k)
ddir(x, k)
}
