% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkewNorm.R
\name{dSkewNorm}
\alias{dSkewNorm}
\title{Density function for Generalized Normal distribution with skew parameter.}
\usage{
dSkewNorm(x, q1, q3, s = 0)
}
\arguments{
\item{x}{vector of quantiles. If `length(n)>1``, the length is taken to be the number required.}

\item{q1}{minimum value}

\item{q3}{maximum value}

\item{s}{skew parameter betweem -1 and 1. When skew parameter equal 0, distribution is normal.}
}
\value{
a vector of probabilities of length equal to \code{length(x)}.
}
\description{
Density function for Generalized Normal distribution with skew parameter.
}
\examples{
dSkewNorm(12, 10, 20, 0.5)
}
