% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elicit.R
\name{fit_dir}
\alias{fit_dir}
\alias{fit_gendir}
\title{Fit Dirichlet or Generalized Dirichlet distribution to the dataframe of elicited conditional QPPs}
\usage{
fit_dir(df, id_col, spt_col, prb_col)

fit_gendir(df, id_col, spt_col, prb_col)
}
\arguments{
\item{df}{data frame of conditional QPPs with 3 columns: category id, SPTs (probability triplets per category), values (probability value for Qdirichlet)}

\item{id_col}{character name of the column containing category id}

\item{spt_col}{character name of the column containing pribability triplets per category}

\item{prb_col}{character name of the column containing values (probabilities for QDirichlet)}
}
\value{
data frame with 2(3) columns containing category id and the parameters of Dirichlet or
Generalized Dirichlet distribution
}
\description{
Fit Dirichlet or Generalized Dirichlet distribution to the dataframe of elicited conditional QPPs
}
\examples{
df <- data.frame(id=rep(1:3, each=3),
                 spt=rep(0.25*1:3, 3),
                 values=c(7, 9, 12, 34, 41, 50, 5, 7.5, 15)/100)
 fit_dir(df, "id", "spt", "values")
 fit_gendir(df, "id", "spt", "values")
}
