% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.R
\name{fexp}
\alias{fexp}
\alias{dqexp}
\alias{ffexp}
\title{Quantile derivative functions for exponential distributions}
\usage{
fexp(u, lambda)

dqexp(u, lambda)

ffexp(u, lambda)
}
\arguments{
\item{u}{numeric probability vector}

\item{lambda}{parameter of exponential distribution}
}
\value{
quantile density, density quantile or quantile convexity for \code{fexp}, \code{dqexp} and \code{ffexp}, respectively
}
\description{
Quantile derivative functions for exponential distributions
}
\examples{
fexp(0.1, 0.5)
dqexp(0.1, 0.5)
ffexp(0.1, 0.5)
}
