% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{fit_beta}
\alias{fit_beta}
\title{Fit beta distribution to symmetrical quantile triplet of probabilities}
\usage{
fit_beta(lq, mq, uq, alpha = 0.25, ...)
}
\arguments{
\item{lq, mq, uq}{values of lower, median and upper quantiles. All values should be between 0 and 1.}

\item{alpha}{share of density in each tail. Default is 0.25, which means that \code{lq} corresponds to 25th quantile and \code{uq} to 75th quantile.}

\item{...}{other arguments passed to \code{stats::optim}}
}
\value{
data.frame with one row and two columns: \code{alpha} and \code{beta}
}
\description{
Fit beta distribution to symmetrical quantile triplet of probabilities
}
\examples{
fit_beta(0.6, 0.65, 0.71)
}
