% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fld.R
\name{fld}
\alias{fld}
\alias{qfld}
\alias{ffld}
\alias{dqfld}
\alias{pfld}
\title{Flattened Generalized Logistic Distribution (FLD)}
\usage{
qfld(p, bt, dlt, k)

ffld(p, bt, dlt, k)

dqfld(p, bt, dlt, k)

pfld(q, bt, dlt, k, tol = 1e-06)
}
\arguments{
\item{p}{vector of probabilities}

\item{bt}{parameter beta of FGLD distribution}

\item{dlt}{parameter delta of FGLD distribution}

\item{k}{parameter k of FGLD distribution}

\item{q}{vector of quantiles}

\item{tol}{tolerance value for optimization. Default value 1e-06}
}
\value{
vector
}
\description{
Defines quantile function (ICDF), probability density and quantile density function as well as probability function (CDF) for flattened generalized logistic distribution
}
\examples{
qfld(0.1, 0.5, 0.3, 1)
}
