% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsld.R
\name{fsld}
\alias{fsld}
\alias{qfsld}
\alias{ffsld}
\alias{dqfsld}
\alias{rfsld}
\alias{pfsld}
\title{Flattened Skew-Logistic Distribution (FSLD)}
\usage{
qfsld(p, bt, k, dlt = 0.5, a = 0)

ffsld(p, bt, k, dlt = 0.5, a = 0)

dqfsld(p, bt, k, dlt = 0.5, a = 0, log = FALSE)

rfsld(n, bt, k, dlt = 0.5, a = 0)

pfsld(q, ..., lower = 0, upper = 1, tol = 1e-06, silent = TRUE, trace = 0)
}
\arguments{
\item{p}{vector of probabilities}

\item{bt}{parameter beta of GFLD distribution (scale). Should be non-negative.}

\item{k}{parameter k of GFLD distribution (shape). Should be non-negative.}

\item{dlt}{parameter delta of GFLD distribution(mixing parameter). Should be within the interval (0,1), default is 0.5}

\item{a}{location parameter alpha of GFLD distribution(location parameter), default is 0}

\item{log}{logical; if TRUE, log density is returnes. Default is FALSE}

\item{n}{numeric; number of samples to draw from FLD distribution}

\item{q}{vector of quantiles}

\item{...}{used by method}

\item{lower, upper}{the \code{stats::uniroot} lower and upper end points of the interval to be searched. Defaults are 0 and 1, respectively}

\item{tol}{the \code{stats::uniroot} desired accuracy (convergence tolerance). Default value 1e-06}

\item{silent}{the \code{base::try} argument. Default is TRUE}

\item{trace}{integer number passed to \code{stats::uniroot}; if positive, tracing information is produced. Higher values giving more details.}
}
\value{
vector
}
\description{
Defines quantile function (ICDF), probability density and quantile density function as well as probability function (CDF) for generalized flattened logistic distribution
}
\examples{
qfsld(0.1, 0.5, 0.3, 0.5)
# centered gfld
p <- runif(1e4)
x <- qfsld(p, 0.25, 1)-qfsld(0.5, 0.25, 1)
hist(x,30)
}
