% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{dgendir}
\alias{dgendir}
\alias{rgendir}
\title{Generalized Dirichlet (Connor-Mosimann) distribution}
\usage{
dgendir(x, a, b, log = FALSE)

rgendir(n, a, b)
}
\arguments{
\item{x}{simplex vector x of length n}

\item{a}{parameter vector a of length n-1}

\item{b}{parameter vector b of length n-1}

\item{log}{logical indicating whether log density should be returned instead. Default FALSE}

\item{n}{number of observations to draw}
}
\description{
Generalized Dirichlet (Connor-Mosimann) distribution
}
\examples{
x <- c(0.15, 0.40, 0.05, 0.40)
a <- c(4.6, 6.7, 1.32)
b <- c(39.4, 7.77, 5.10)
dgendir(x,a,b)
rgendir(1,a,b)
}
