% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genexp.R
\name{pgenexp}
\alias{pgenexp}
\alias{dgenexp}
\alias{qgenexp}
\alias{fgenexp}
\alias{rgenexp}
\title{Generalized exponential distribution}
\usage{
pgenexp(x, alpha, lambda)

dgenexp(x, alpha, lambda)

qgenexp(p, alpha, lambda)

fgenexp(p, alpha, lambda)

rgenexp(n, alpha, lambda)
}
\arguments{
\item{x}{numeric vector of values}

\item{alpha}{parameter of generalized exponential distribution}

\item{lambda}{parameter of generalized exponential distribution}

\item{p}{numeric vector of cumulative probabilities}

\item{n}{number of samples to draw}
}
\value{
cum
}
\description{
Generalized exponential distribution with CDF F(x)=(1-exp(-lambda*x))^alpha
}
\examples{
pgenexp(1, 1, 1)
}
