% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld.R
\name{qgld}
\alias{qgld}
\alias{fgld}
\alias{dqgld}
\alias{rgld}
\alias{pgld}
\alias{is_gld_valid}
\title{Generalized Lambda Distribution (GLD)}
\usage{
qgld(u, l1, l2, l3, l4)

fgld(u, l1, l2, l3, l4, log = FALSE)

dqgld(u, l1, l2, l3, l4, log = FALSE)

rgld(n, l1, l2, l3, l4)

pgld(q, l1, l2, l3, l4, n_grid = 50L, s_grid = 5L, tol = 1e-15, maxiter = 1000)

is_gld_valid(l1, l2, l3, l4, n_grid = 100L, s_grid = 2L)
}
\arguments{
\item{u}{numeric vector of probabilities}

\item{l1}{GLD parameter \eqn{\lambda_1}}

\item{l2}{GLD parameter \eqn{\lambda_2}}

\item{l3}{GLD parameter \eqn{\lambda_3}}

\item{l4}{GLD parameter \eqn{\lambda_4}}

\item{log}{should the log density be returned. Default=FALSE}

\item{n}{numeric number of samples to draw}

\item{q}{quantile value for which the corresponding cumulative probability value should be found.}

\item{n_grid}{integer size of helper grid to be passed to \code{make_pgrid}. Default is 50}

\item{s_grid}{integer beta shape of helper grid to be passed to \code{make_pgrid}. Default is 5}

\item{tol}{numeric tolerance balue for approximating CDF. Default 1e-15}

\item{maxiter}{numeric maximum number of iteration}
}
\value{
quantiles, QDF, DQF, random samples or probabilities of GLD
}
\description{
Generalized Lambda Distribution (GLD)
}
\examples{
p_grd <- make_pgrid()
qgld(p_grd, 1, -1, -1/8, -1/32)
}
