% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gldcsw.R
\name{qGLDcsw}
\alias{qGLDcsw}
\alias{fGLDcsw}
\alias{dqGLDcsw}
\alias{rGLDcsw}
\alias{pGLDcsw}
\title{Generalized Lambda Distribution (GLD) CSW parameterization}
\usage{
qGLDcsw(u, mu, sg, chi, xi)

fGLDcsw(u, mu, sg, chi, xi)

dqGLDcsw(u, mu, sg, chi, xi, log = FALSE)

rGLDcsw(n, mu, sg, chi, xi)

pGLDcsw(q, ..., lower = 0, upper = 1, tol = 1e-06, silent = TRUE, trace = 0)
}
\arguments{
\item{u}{numeric vector of probabilities}

\item{mu}{CSW GLD median parameter \eqn{\mu}}

\item{sg}{CSW GLD interquartile range parameter \eqn{\sg}}

\item{chi}{CSW GLD assymetry parameter \eqn{\chi}}

\item{xi}{CSW GLD steepness parameter \eqn{\lambda_4}}

\item{log}{should the log density be returned. Default=FALSE}

\item{n}{numeric number of samples to draw}

\item{q}{vector of quantiles}

\item{...}{used by method}

\item{lower, upper}{the \code{stats::uniroot} lower and upper end points of the interval to be searched. Defaults are 0 and 1, respectively}

\item{tol}{the \code{stats::uniroot} desired accuracy (convergence tolerance). Default value 1e-06}

\item{silent}{the \code{base::try} argument. Default is TRUE}

\item{trace}{integer number passed to \code{stats::uniroot}; if positive, tracing information is produced. Higher values giving more details.}
}
\value{
quantiles, QDF, DQF, random samples or probabilities of GLD (CSW parameterization)
}
\description{
Quantile function, quantile density, density quantile and inverse quantile
functions for GLD distribution with CSW parameterization.
\code{aGLDcsw_mean} and \code{aGLDcsw_median} are theoretical mean and median, which can be used
for adjusting the quantile likelihood.
}
\examples{
p_grd <- make_pgrid()
qGLDcsw(p_grd, 1, 1, -1/8, 1/32)
}
