% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gldrs.R
\name{qGLDrs}
\alias{qGLDrs}
\alias{fGLDrs}
\alias{dqGLDrs}
\alias{rGLDrs}
\alias{pGLDrs}
\alias{is_GLDrs_valid}
\alias{aGLDrs_mean}
\alias{aGLDrs_median}
\title{Generalized Lambda Distribution (GLD) RS parameterization}
\usage{
qGLDrs(u, l1, l2, l3, l4)

fGLDrs(u, l1, l2, l3, l4)

dqGLDrs(u, l1, l2, l3, l4, log = FALSE)

rGLDrs(n, l1, l2, l3, l4)

pGLDrs(q, ..., lower = 0, upper = 1, tol = 1e-06, silent = TRUE, trace = 0)

is_GLDrs_valid(l1, l2, l3, l4)

aGLDrs_mean(l1, l2, l3, l4)

aGLDrs_median(l1, l2, l3, l4)
}
\arguments{
\item{u}{numeric vector of probabilities}

\item{l1}{GLD parameter \eqn{\lambda_1}}

\item{l2}{GLD parameter \eqn{\lambda_2}}

\item{l3}{GLD parameter \eqn{\lambda_3}}

\item{l4}{GLD parameter \eqn{\lambda_4}}

\item{log}{should the log density be returned. Default=FALSE}

\item{n}{numeric number of samples to draw}

\item{q}{vector of quantiles}

\item{...}{used by method}

\item{lower, upper}{the \code{stats::uniroot} lower and upper end points of the interval to be searched. Defaults are 0 and 1, respectively}

\item{tol}{the \code{stats::uniroot} desired accuracy (convergence tolerance). Default value 1e-06}

\item{silent}{the \code{base::try} argument. Default is TRUE}

\item{trace}{integer number passed to \code{stats::uniroot}; if positive, tracing information is produced. Higher values giving more details.}
}
\value{
quantiles, QDF, DQF, random samples or probabilities of GLD
}
\description{
Quantile function, quantile density, density quantile and inverse quantile
functions for GLD distribution with RS parameterization.
\code{aGLDrs_mean} and \code{aGLDrs_median} are theoretical mean and median, which can be used
for adjusting the quantile likelihood.
}
\examples{
p_grd <- make_pgrid()
is_GLDrs_valid(1, -1, -1/8, -1/32)
qGLDrs(p_grd, 1, -1, -1/8, -1/32)
}
