% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnh.R
\name{qgnh}
\alias{qgnh}
\alias{fgnh}
\alias{dqgnh}
\alias{rgnh}
\alias{pgnh}
\alias{is_gnh_valid}
\title{Functions for g-and-h distribution}
\usage{
qgnh(p, A, B, C = 0.8, g, h, zscale = FALSE)

fgnh(p, A, B, C = 0.8, g, h, log = FALSE, zscale = FALSE)

dqgnh(p, A, B, C = 0.8, g, h, log = FALSE, zscale = FALSE)

rgnh(n, A, B, C = 0.8, g, h)

pgnh(
  q,
  A,
  B,
  C = 0.8,
  g,
  h,
  n_grid = 100L,
  s_grid = 5L,
  tol = 1e-15,
  maxiter = 1000
)

is_gnh_valid(A, B, C = 0.8, g, h)
}
\arguments{
\item{p}{numeric vector of probabilities}

\item{A}{numeric location parameter of g-and-h distribution}

\item{B}{positive numeric scale parameter of g-and-h distribution}

\item{C}{numeric parameter of g-and-h distribution. Default is 0.8}

\item{g}{numeric skeweness parameter of g-and-h distribution}

\item{h}{non-negative numeric kurtosis parameter of g-and-h distribution}

\item{zscale}{is the probability input provided on scale of \code{z} values, i.e. as \code{z=qnorm(p)}. Default is FALSE}

\item{log}{logical should the result be returned as log(). Default is FALSE}

\item{n}{numeric number of samples to draw}

\item{q}{quantile value for which the corresponding cumulative probability value should be found.}

\item{n_grid}{integer size of helper grid to be passed to \code{make_pgrid}. Default is 50}

\item{s_grid}{integer beta shape of helper grid to be passed to \code{make_pgrid}. Default is 5}

\item{tol}{numeric tolerance balue for approximating CDF. Default 1e-15}

\item{maxiter}{numeric maximum number of iteration}
}
\value{
vector of quantiles for g-and-h distribution

vector of quantiles for g-and-h distribution
}
\description{
Functions for g-and-h distribution
}
\examples{
qgnh(0.1, 3,1,0.8, 2,0.5)
}
