% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnk.R
\name{qgnk}
\alias{qgnk}
\alias{fgnk}
\alias{dqgnk}
\alias{rgnk}
\alias{pgnk}
\alias{is_gnk_valid}
\title{Functions for g-and-k distribution}
\usage{
qgnk(p, A, B, C = 0.8, g, k, zscale = FALSE)

fgnk(p, A, B, C = 0.8, g, k, log = FALSE, zscale = FALSE)

dqgnk(p, A, B, C = 0.8, g, k, log = FALSE, zscale = FALSE)

rgnk(n, A, B, C = 0.8, g, k)

pgnk(q, ..., lower = 0, upper = 1, tol = 1e-06, silent = TRUE, trace = 0)

is_gnk_valid(A, B, C = 0.8, g, k, n_grid = 100L, s_grid = 2L)
}
\arguments{
\item{p}{numeric vector of probabilities}

\item{A}{numeric location parameter of g-and-k distribution}

\item{B}{positive numeric scale parameter of g-and-k distribution}

\item{C}{numeric parameter of g-and-k distribution. Default is 0.8}

\item{g}{numeric skeweness parameter of g-and-k distribution}

\item{k}{non-negative numeric kurtosis parameter of g-and-k distribution}

\item{zscale}{is the probability input provided on scale of \code{z} values, i.e. as \code{z=qnorm(p)}. Default is FALSE}

\item{log}{logical should the result be returned as log(). Default is FALSE}

\item{n}{numeric number of samples to draw}

\item{q}{vector of quantiles}

\item{...}{used by method}

\item{lower, upper}{the \code{stats::uniroot} lower and upper end points of the interval to be searched. Defaults are 0 and 1, respectively}

\item{tol}{the \code{stats::uniroot} desired accuracy (convergence tolerance). Default value 1e-06}

\item{silent}{the \code{base::try} argument. Default is TRUE}

\item{trace}{integer number passed to \code{stats::uniroot}; if positive, tracing information is produced. Higher values giving more details.}

\item{n_grid}{integer size of helper grid to be passed to \code{make_pgrid}. Default is 50}

\item{s_grid}{integer beta shape of helper grid to be passed to \code{make_pgrid}. Default is 5}
}
\value{
vector of quantiles for g-and-k distribution
}
\description{
Functions for g-and-k distribution
}
\examples{
qgnk(0.1, 3,1,0.8, 2,0.5)
is_gnk_valid(A=5, B=5, C=0.8, g=0.5, k=-0.3) #FALSE
is_gnk_valid(A=5, B=5, C=0.8, g=0.5, k=-0.1) #TRUE
}
