% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govindarajulu.R
\name{govindarajulu}
\alias{govindarajulu}
\alias{qgovindarajulu}
\alias{fgovindarajulu}
\alias{dqgovindarajulu}
\alias{pgovindarajulu}
\alias{dgovindarajulu}
\title{Govindarajulu probability distribution}
\usage{
qgovindarajulu(p, sg, bt)

fgovindarajulu(p, sg, bt)

dqgovindarajulu(p, sg, bt, log = FALSE)

pgovindarajulu(q, sg, bt, tol = 1e-06)

dgovindarajulu(x, sg, bt, tol = 1e-06)
}
\arguments{
\item{p}{vector of probabilities}

\item{sg}{parameter sigma of Govindarajulu distribution}

\item{bt}{parameter beta of Govindarajulu distribution}

\item{log}{logical; if TRUE, log density is returned. Default is FALSE}

\item{q}{vector of quantiles}

\item{tol}{tolerance value for optimization. Default value 1e-06}

\item{x}{numeric vector}
}
\value{
vector
}
\description{
Defines quantile function (ICDF), probability density and quantile density function as well as probability function (CDF) for Govindarajulu distribution
}
\examples{
qgovindarajulu(0.1, 1, 5)
}
