% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JQPD.R
\name{GQPDB}
\alias{GQPDB}
\alias{qGQPDB}
\alias{rGQPDB}
\title{The  G-QPD-B distribution}
\usage{
qGQPDB(
  p,
  q1,
  q2,
  q3,
  lower,
  upper,
  alpha = 0.1,
  qf = stats::qnorm,
  pf = stats::pnorm
)

rGQPDB(
  N,
  q1,
  q2,
  q3,
  lower,
  upper,
  alpha = 0.1,
  qf = stats::qnorm,
  pf = stats::pnorm
)
}
\arguments{
\item{p}{vector of probabilities}

\item{q1, q2, q3}{vectors of values, corresponding to lower, median and upper (symmetrical) quantiles}

\item{lower, upper}{vectors of lower and upper bounds of distribution}

\item{alpha}{vector of proportions of probability density under the lower bound
(or above the upper bound, since the quantiles are symmetrical)}

\item{qf}{quantile function}

\item{pf}{distribution function}

\item{N}{number of samples to draw}
}
\value{
vector of values
}
\description{
Quantile function and random generation for the
Generalized Quantile Parametrized (Bounded) distribution parametrized by three symmetrical
quantiles (\code{q1}, \code{q2} and \code{q3}) and an \code{alpha} argument, representing the proportion of
density below the bottom quantile. All functions are vectorized.
}
\details{
The distribution is created by applying the inverse probit transform T(x)=Phi(x) to the Johnson SU distribution
}
