% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JQPD.R
\name{GQPDS}
\alias{GQPDS}
\alias{qGQPDS}
\alias{rGQPDS}
\title{The Generalized QPD-S distribution}
\usage{
qGQPDS(p, q1, q2, q3, lower = 0, alpha = 0.1, qf = stats::qnorm)

rGQPDS(N, q1, q2, q3, lower = 0, alpha = 0.1, qf = stats::qnorm)
}
\arguments{
\item{p}{vector of probabilities}

\item{q1, q2, q3}{vectors of values, corresponding to lower, median and upper (symmetrical) quantiles}

\item{lower}{vector of lower bounds of distribution}

\item{alpha}{vector of proportions of probability density under the lower bound
(or above the upper bound, since the quantiles are symmetrical)}

\item{qf}{bare symmetrical unbounded standard quantile function}

\item{N}{number of samples to draw from GQPDS}
}
\value{
vector of values
}
\description{
Quantile function and random generation for the
Generalized Quantile Parametrized (Semibounded) distribution parametrized by three symmetrical
quantiles (\code{q1}, \code{q2} and \code{q3}) and an \code{alpha} argument, representing the proportion of
density below the bottom quantile. All functions are vectorized.
}
\details{
The distribution is created by applying the exponential transform T(x)=exp(x) to the Johnson SU distribution
}
\examples{
qGQPDS(0.6, 20,50,90)
}
