% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isim.R
\name{isim_to_cprob}
\alias{isim_to_cprob}
\title{Convert index simplex to cumulative probability}
\usage{
isim_to_cprob(s, idx)
}
\arguments{
\item{s}{vector or matrix of simplices}

\item{idx}{integer vector of indices}
}
\value{
matrix of cumulative probabilities
}
\description{
Convert index simplex to cumulative probability
}
\examples{
s <- c(0.1, 0.4, 0.15, 0.35)
ss <- matrix(c(0.1, 0.4, 0.15, 0.35,
               0.13, 0.26, 0.05, 0.56),
               ncol=4, byrow=TRUE)
idx <- c(1L, 2L, 4L, 3L)
isim_to_cprob(s, idx)
isim_to_cprob(ss, idx)
}
