% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JQPD.R
\name{JQPDB}
\alias{JQPDB}
\alias{qJQPDB}
\alias{pJQPDB}
\alias{dJQPDB}
\title{The J-QPD-B distribution}
\usage{
qJQPDB(p, q1, q2, q3, lower, upper, alpha = 0.1)

pJQPDB(q, q1, q2, q3, lower, upper, alpha = 0.1)

dJQPDB(x, q1, q2, q3, lower, upper, alpha = 0.1)
}
\arguments{
\item{p}{vector of probabilities}

\item{q1, q2, q3}{vectors of values, corresponding to lower, median and upper (symmetrical) quantiles}

\item{lower, upper}{vectors of lower and upper bounds of distribution}

\item{alpha}{vector of proportions of probability density under the lower bound
(or above the upper bound, since the quantiles are symmetrical)}

\item{x, q}{vector of quantiles}
}
\value{
vector of values
}
\description{
Density, distribution function, quantile function and random generation for the
Johnson Quantile Parametrized (Bounded) distribution parametrized by three symmetrical
quantiles (\code{q1}, \code{q2} and \code{q3}) and an \code{alpha} argument, representing the proportion of
density below the bottom quantile. All functions are vectorized.
}
\examples{
# should result in c(0,1,5,12,20)
qJQPDB(c(0, 0.05, 0.5, 0.95, 1), 1, 5, 12, 0, 20, alpha=0.05)
# should return c(0.00, 0.05, 0.50, 0.95, 1.00)
pJQPDB(c(0, 1, 5, 12, 20), 1,5,12, 0, 20, alpha=0.05)
# should return vector with first and last element equal to NaN
dJQPDB(c(0, 1, 5, 12, 20), 1,5,12, 0, 20, alpha=0.05)
}
