% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JQPD.R
\name{JQPDS}
\alias{JQPDS}
\alias{qJQPDS}
\alias{pJQPDS}
\alias{dJQPDS}
\alias{rJQPDS}
\title{The J-QPD-S distribution}
\usage{
qJQPDS(p, q1, q2, q3, lower = 0, alpha = 0.1)

pJQPDS(q, q1, q2, q3, lower = 0, alpha = 0.1)

dJQPDS(x, q1, q2, q3, lower = 0, alpha = 0.1)

rJQPDS(N, q1, q2, q3, lower = 0, alpha = 0.1)
}
\arguments{
\item{p}{vector of probabilities}

\item{q1, q2, q3}{vectors of values, corresponding to lower, median and upper (symmetrical) quantiles}

\item{lower}{vector of lower bounds of distribution}

\item{alpha}{vector of proportions of probability density under the lower bound
(or above the upper bound, since the quantiles are symmetrical)}

\item{q}{vector of quantiles}

\item{x}{vector of observations}

\item{N}{number of samples for draw}
}
\value{
vector of values
}
\description{
Density, distribution function, quantile function and random generation for the
Johnson Quantile Parametrized (Semibounded) distribution parametrized by three symmetrical
quantiles (\code{q1}, \code{q2} and \code{q3}) and an \code{alpha} argument, representing the proportion of
density below the bottom quantile. All functions are vectorized.
}
\details{
The distribution is created by applying the exponential transform T(x)=exp(x) to the Johnson SU distribution
}
\examples{
# should result in c(0,1,5,12,20)
qJQPDB(c(0, 0.05, 0.5, 0.95, 1), 1, 5, 12, 0, 20, alpha=0.05)
qJQPDS(0.6, 20,50,90)
# should return c(0.00, 0.05, 0.50, 0.95, 1.00)
pJQPDB(c(0, 1, 5, 12, 20), 1,5,12, 0, 20, alpha=0.05)
# should return vector with first and last element equal to NaN
dJQPDB(c(0, 1, 5, 12, 20), 1,5,12, 0, 20, alpha=0.05)
# should return vector with first and last element equal to NaN
}
