% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JQPD.R
\name{JQPDS2}
\alias{JQPDS2}
\alias{qJQPDS2}
\alias{pJQPDS2}
\alias{dJQPDS2}
\title{The J-QPD-S Type II distribution}
\usage{
qJQPDS2(p, q1, q2, q3, lower = 0, alpha = 0.1)

pJQPDS2(q, q1, q2, q3, lower = 0, alpha = 0.1)

dJQPDS2(x, q1, q2, q3, lower = 0, alpha = 0.1)
}
\arguments{
\item{p}{vector of probabilities}

\item{q1, q2, q3}{vectors of values, corresponding to lower, median and upper (symmetrical) quantiles}

\item{lower}{vector of lower bounds of distribution}

\item{alpha}{vector of proportions of probability density under the lower bound
(or above the upper bound, since the quantiles are symmetrical)}

\item{q}{vector of quantiles}

\item{x}{vector of observations}
}
\value{
vector of values
}
\description{
Density, distribution function, quantile function and random generation for the
Johnson Quantile Parametrized (Semibounded) Type II distribution parametrized by three symmetrical
quantiles (\code{q1}, \code{q2} and \code{q3}) and an \code{alpha} argument, representing the proportion of
density below the bottom quantile. All functions are vectorized.
}
\details{
The distribution is created by applying the exponential transform T(x)=exp(x) to the Johnson SU distribution
}
\examples{

qJQPDS2(0.6, 20,50,90)
pJQPDS2(15, 20,50,90)
# should return vector with first and last element equal to NaN
dJQPDS2(15, 20,50,90)
}
