% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logis.R
\name{flogis}
\alias{flogis}
\alias{dqlogis}
\title{Derivative quantile functions for logistic distribution}
\usage{
flogis(p, location = 0, scale = 1, log = FALSE)

dqlogis(p, location = 0, scale = 1, log = FALSE)
}
\arguments{
\item{p}{numeric probability}

\item{location}{numeric parameter of logistic distribution. Default is 0}

\item{scale}{numeric parameter of logistic distribution. Default is 1}

\item{log}{logical, should the result be returned as a log. Default is FALSE}
}
\value{
returns QDF (\code{flogis}) and DQF (\code{dqlogis}) of logistic distribution
}
\description{
Calculate QDF \code{flogis} and DQF \code{dqlogis} for the logistic distribution
}
\examples{
p_grd <- make_pgrid()
flogis(p_grd)
dqlogis(p_grd)
}
