% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myerson.R
\name{logitMyerson}
\alias{logitMyerson}
\alias{qlogitMyerson}
\alias{flogitMyerson}
\alias{dqlogitMyerson}
\alias{rlogitMyerson}
\alias{plogitMyerson}
\alias{dlogitMyerson}
\title{Density function for logit Myerson distribution (Myerson distribution with logit base function).}
\usage{
qlogitMyerson(p, q1, q2, q3, alpha = 0.25, lower.tail = TRUE, log.p = FALSE)

flogitMyerson(
  p,
  q1,
  q2,
  q3,
  alpha = 0.25,
  log = FALSE,
  lower.tail = TRUE,
  log.p = FALSE
)

dqlogitMyerson(
  p,
  q1,
  q2,
  q3,
  alpha = 0.25,
  log = FALSE,
  lower.tail = TRUE,
  log.p = FALSE
)

rlogitMyerson(n, q1, q2, q3, alpha = 0.25, lower.tail = TRUE, log.p = FALSE)

plogitMyerson(q, q1, q2, q3, alpha = 0.25, lower.tail = TRUE, log.p = FALSE)

dlogitMyerson(x, q1, q2, q3, alpha = 0.25, log = FALSE)
}
\arguments{
\item{p}{vector of probabilities}

\item{q1, q2, q3}{numeric values representing bottom quantile, middle quantile (50th percentile)
and top quantile. Quantiles are assumed to be symmetrical.}

\item{alpha}{numerical fixed proportion of distribution below the bottom quantile.
Default value is 0.25}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X\leq x]}, otherwise, \eqn{P[X > x]}}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p)}

\item{n}{number of observations. If `length(n)>1``, the length is taken to be the number required.}

\item{x, q}{vector of quantiles.}
}
\value{
a vector of probabilities of length equal to \code{length(x)}

a vector of exceedance probabilities of length equal to \code{length(x)}.
}
\description{
Density function for logit Myerson distribution (Myerson distribution with logit base function).
}
\examples{
qlogitMyerson(0.25, 10, 20, 40)
flogitMyerson(0.25, 10, 20, 40)
dqlogitMyerson(0.25, 10, 20, 40)
rlogitMyerson(1, 10, 20, 40)
plogitMyerson(0.25, 10, 20, 40)
dlogitMyerson(15, 10, 20, 40)
}
