% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebyshev.R
\name{m11_to_ab}
\alias{m11_to_ab}
\alias{ab_to_m11}
\title{Transform values from any interval to [-1, 1] and back}
\usage{
m11_to_ab(z, from = 0, to = 1)

ab_to_m11(x, from = 0, to = 1)
}
\arguments{
\item{from, to}{the interval(s) to which x values belong. Defaults are from=0,to=1}

\item{x, z}{vector of values on regular and [-1,1] scale, respectively}
}
\value{
returns values on the other scale
}
\description{
Transform values from any interval to [-1, 1] and back
}
\examples{
m11_to_ab(c(-0.5, 0, 0.5))
ab_to_m11(c(0.25, 0.5, 0.75))
}
