% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{make_pgrid}
\alias{make_pgrid}
\alias{make_tgrid}
\title{Make probability grid}
\usage{
make_pgrid(n = 50L, s = 5L, trim = TRUE)

make_tgrid(n = 50L, tier = 3L, tail = 0.1)
}
\arguments{
\item{n}{integer length of the grid. Default is 50}

\item{s}{beta distribution shape parameter, passed to both \code{shape1} and \code{shape2} of \code{pbeta()}. For uniform grid choose 1. Default is 5.}

\item{trim}{logical, should the 0 and 1 (tails of the grid) be trimmed. Default is \code{TRUE}}

\item{tier}{integer number of tiers in the linear grid. Each tier contains \code{tail} share of the previous tier. Default is 3.}

\item{tail}{real number representing share of grid in each tier}
}
\value{
probability grid vector of length n
}
\description{
Functions for creating a probability grid using different method
\code{make_pgrid} uses beta-distribution method
\code{make_tgrid} uses tiered linear method
}
\examples{
make_pgrid(100)
make_pgrid(100, 1, FALSE) #uniform grid including 0 and 1
make_tgrid(100,3,0.1)
}
