% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_pgrid}
\alias{make_pgrid}
\title{Make probability grid using beta-distribution method}
\usage{
make_pgrid(n = 100, s = 10, trim = TRUE)
}
\arguments{
\item{n}{integer length of the grid. Default is 100}

\item{s}{beta distribution shape parameter. For uniform grid choose 1. Default is 10.}

\item{trim}{logical, should the 0 and 1 (tails of the grid) be trimmed. Default is \code{TRUE}}
}
\value{
probability grid vector of length n
}
\description{
Make probability grid using beta-distribution method
}
\examples{
make_pgrid(100)
make_pgrid(100, 1, FALSE) #uniform grid including 0 and 1
}
