% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalog.R
\name{fit_metalog}
\alias{fit_metalog}
\alias{qmetalog}
\alias{fmetalog}
\alias{pmetalog}
\alias{rmetalog}
\alias{is_metalog_valid}
\title{Metalog distribution functions}
\usage{
fit_metalog(p, q, bl = NULL, bu = NULL)

qmetalog(p, a, bl = NULL, bu = NULL)

fmetalog(p, a, bl = NULL, bu = NULL)

pmetalog(x, a, bl = NULL, bu = NULL, tol = 1e-06, maxiter = 1e+06)

rmetalog(n, a, bl = NULL, bu = NULL)

is_metalog_valid(a, bl = NULL, bu = NULL)
}
\arguments{
\item{p}{vector of cumulative probabilities}

\item{q}{vector of quantile values corresponding to cumulative probabilities \code{p}}

\item{bl}{real value of lower boundary (for bounded metalog). Default NULL}

\item{bu}{real value of upper boundary (for bounded metalog). Default NULL}

\item{a}{vector of \code{a}-coefficient parameters of metalog distribution}

\item{x}{real vector of values}

\item{tol}{tolerance value, default is 1e-6}

\item{maxiter}{maximum number of iterations for approximation, default is 1e6}

\item{n}{integer value correponding to number of samples to draw}
}
\description{
Functions for fitting and sampling from metalog distribution
}
\details{
\code{fit_metalog} is for fitting the metalog function to the set of QP values.
Number of metalog terms will match the number of QP pairs.``
\code{qmetalog} is a quantile function.
`fmetalog` is a quantile density function q(u). The reciprocal of it is density quantile function f(Q(p)).
`pmetalog` is an approximation of the cumulative density function.
`rmetalog` is an RNG.
`is_metalog_valid` is a function for checking if the metalog is valid
}
\examples{
p <- c(0.1, 0.5, 0.9)
q <- c(4, 9, 12)
fit_metalog(p,q)
a <- c(9,  1.8, -1.13)
qmetalog(p, a)
fmetalog(p, a)
x <- c(5, 9, 14)
pmetalog(x, a)
rmetalog(100, a)
is_metalog_valid(a)
}
