% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myerson.R
\name{MMyerson}
\alias{MMyerson}
\alias{qMMyerson}
\title{Quantile function for Modified Myerson distribution.}
\usage{
qMMyerson(p, q1, q2, q3, alpha = 0.25, sfun = stats::qnorm)
}
\arguments{
\item{p}{vector of probabilities}

\item{q1, q2, q3}{numeric values representing bottom quantile, middle quantile (50th percentile)
and top quantile. Quantiles are assumed to be symmetrical.}

\item{alpha}{numerical fixed proportion of distribution below the bottom quantile.
Default value is 0.25}

\item{sfun}{function; standard vectorized quantile function with a single depth parameter}
}
\value{
a vector of quantilesof length equal to \code{length(x)}
}
\description{
Quantile function for Modified Myerson distribution.
}
\examples{
pgrd <- make_pgrid()
qMMyerson(pgrd, 10, 20, 40, alpha=0.1, sfun=stats::qlogis)
}
