% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{monopower}
\alias{monopower}
\title{internal function for monotonic power}
\usage{
monopower(x, pow = 1)
}
\arguments{
\item{x}{numeric vector of centered depths i.e. x=u-0.5}

\item{pow}{numeric power to raise the depth into}
}
\value{
monopowered centered depth
}
\description{
internal function for monotonic power
}
\references{
Ahmadabadi MN, Farjami Y, Moghadam MB. 2012. Approximating Distributions by Extended Generalized Lambda Distribution (XGLD). Communications in Statistics - Simulation and Computation. 41(1):1–23. https://doi.org/10.1080/03610911003681503
}
\keyword{internal}
