% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{fnorm}
\alias{fnorm}
\alias{dqnorm}
\title{Derivative quantile functions for normal distribution}
\usage{
fnorm(p, mean = 0, sd = 1, log = FALSE, log.p = FALSE)

dqnorm(p, mean = 0, sd = 1, log = FALSE, log.p = FALSE)
}
\arguments{
\item{p}{numeric probability}

\item{mean}{numeric parameter of normal distribution. Default is 0}

\item{sd}{numeric parameter of normal distribution. Default is 1}

\item{log}{logical, should the result be returned as a log. Default is FALSE}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p). Default is FALSE}
}
\value{
returns QDF (\code{fnorm}) and DQF (\code{dqnorm}) of normal distribution
}
\description{
Calculate QDF \code{fnorm} and DQF \code{dqnorm} for the normal distribution
}
\examples{
p_grd <- make_pgrid()
fnorm(p_grd)
dqnorm(p_grd)
}
