% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly.R
\name{eval_poly}
\alias{eval_poly}
\title{Evaluate a polynomial in reverse order}
\usage{
eval_poly(x, b, deriv = FALSE)
}
\arguments{
\item{x}{numeric vector of values to evaluate a polynomial at}

\item{b}{numeric vector of coefficients (highest degree first, degree 0 last)}

\item{deriv}{logical. Should the value of the derivative be returned instead?. Default is FALSE.}
}
\value{
quantiles, QDF, DQF, random samples or probabilities of GLD (CSW parameterization)
}
\description{
Evaluate the polynomial using the Horner scheme
}
\examples{
eval_poly(1:3, c(0.1, 0.2, 0.6, 1.23, 4.43))
}
