% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkewNorm.R
\name{pSkewNorm}
\alias{pSkewNorm}
\title{Cumulative distribution function for  Generalized Normal distribution with skew parameter.}
\usage{
pSkewNorm(x, q1, q3, s = 0)
}
\arguments{
\item{x}{vector of quantiles. If `length(n)>1``, the length is taken to be the number required.}

\item{q1}{minimum value}

\item{q3}{maximum value}

\item{s}{skew parameter betweem -1 and 1. When skew parameter equal 0, distribution is normal.}
}
\value{
a vector of exceedance probabilities of length equal to \code{length(x)}.
}
\description{
Cumulative distribution function for  Generalized Normal distribution with skew parameter.
}
\examples{
pSkewNorm(12, 10, 20, 0.5)
}
