% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JQPD.R
\name{qJDS_SB}
\alias{qJDS_SB}
\title{Quantile function of Johnson Distribution System (JDS) SB distribution
The transformation function is inverse logit T(y)=exp(y)/(1+exp(y))}
\usage{
qJDS_SB(p, zeta, lambda, sigma, gamma)
}
\description{
Quantile function of Johnson Distribution System (JDS) SB distribution
The transformation function is inverse logit T(y)=exp(y)/(1+exp(y))
}
\keyword{internal}
