% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JQPD.R
\name{qJDS_SU}
\alias{qJDS_SU}
\title{Quantile function of Johnson Distribution System (JDS) SU distribution
The transformation function is hyperbolic sine transform}
\usage{
qJDS_SU(p, zeta, lambda, sigma, gamma, qf = stats::qnorm)
}
\description{
Quantile function of Johnson Distribution System (JDS) SU distribution
The transformation function is hyperbolic sine transform
}
\keyword{internal}
