% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmom.R
\name{qdv_eqf}
\alias{qdv_eqf}
\alias{qsk_eqf}
\alias{qkr_eqf}
\title{Unnormalized empirical robust moments.
\deqn{\sigma_r=Q(3/4)-Q(1/4)}
\deqn{s_r=\frac{Q(3/4)+Q(1/4)-2Q(1/2)}{\sigma_r}}
\deqn{\kappa_r=\frac{Q(7/8)-Q(5/8)+Q(3/8)-Q(1/8)}{\sigma_r}}
These are implemented as \code{qdv_eqf}, \code{qsk_eqf}, and \code{qkr_eqf}, respectively. Note that the robust measure of location is the median \eqn{\mu_r=Q(1/2)}.}
\usage{
qdv_eqf(x, type = 5, lambda = 0.25)

qsk_eqf(x, type = 5, lambda = 0.25)

qkr_eqf(x, type = 5, lambda = 0.25, zeta = lambda/2)
}
\arguments{
\item{x}{numerical sample to compute the quantiles from}

\item{type}{parameters, passed to \code{quantile} function}

\item{lambda}{probability < 0.5 corresponding to the tail in the robust deviation. Default is 0.25}

\item{zeta}{probability \code{zeta<lambda} corresponding to the tail in robust kurtosis. Default is \code{lambda/2}}
}
\value{
numeric value of robust moment
}
\description{
Unnormalized empirical robust moments.
\deqn{\sigma_r=Q(3/4)-Q(1/4)}
\deqn{s_r=\frac{Q(3/4)+Q(1/4)-2Q(1/2)}{\sigma_r}}
\deqn{\kappa_r=\frac{Q(7/8)-Q(5/8)+Q(3/8)-Q(1/8)}{\sigma_r}}
These are implemented as \code{qdv_eqf}, \code{qsk_eqf}, and \code{qkr_eqf}, respectively. Note that the robust measure of location is the median \eqn{\mu_r=Q(1/2)}.
}
\examples{

qdv_eqf(1:100) #49.5
qsk_eqf(1:100)
qkr_eqf(1:100)
}
