% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmom.R
\name{qdv_qf}
\alias{qdv_qf}
\alias{qsk_qf}
\alias{qkr_qf}
\title{Robust moments}
\usage{
qdv_qf(.qf, ..., lambda = 0.25)

qsk_qf(.qf, ..., lambda = 0.25)

qkr_qf(.qf, ..., lambda = 0.25, zeta = lambda/2)
}
\arguments{
\item{.qf}{quantile function}

\item{...}{parameters, passed to the of quantile function}

\item{lambda}{probability < 0.5 corresponding to the tail in the robust deviation. Default is 0.25}

\item{zeta}{probability \code{zeta<lambda} corresponding to the tail in robust kurtosis. Default is \code{lambda/2}}
}
\value{
numeric value of robust moment
}
\description{
Unnormalized robust moments.
\deqn{\sigma_r=Q(3/4)-Q(1/4)}
\deqn{s_r=\frac{Q(3/4)+Q(1/4)-2Q(1/2)}{\sigma_r}}
\deqn{\kappa_r=\frac{Q(7/8)-Q(5/8)+Q(3/8)-Q(1/8)}{\sigma_r}}
These are implemented as \code{qdv_qf}, \code{qsk_qf}, and \code{qkr_qf}, respectively. Note that the robust measure of location is the median \eqn{\mu_r=Q(1/2)}.
}
\examples{
qdv_qf(qnorm, 0, 1)
qsk_qf(qnorm, 0, 1)
qkr_qf(qnorm, 0, 1)
}
