% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayleigh.R
\name{drayleigh}
\alias{drayleigh}
\alias{prayleigh}
\alias{qrayleigh}
\alias{rrayleigh}
\alias{frayleigh}
\alias{dqrayleigh}
\alias{ffrayleigh}
\title{Rayleigh distribution}
\usage{
drayleigh(x, sigma)

prayleigh(x, sigma)

qrayleigh(p, sigma)

rrayleigh(n, sigma)

frayleigh(p, sigma)

dqrayleigh(p, sigma, log = FALSE)

ffrayleigh(p, sigma)
}
\arguments{
\item{x}{numeric vector of data}

\item{sigma}{numeric scale parameter of the Rayleigh distribution}

\item{p}{numeric vector of probabilities}

\item{n}{numeric number of random values to draw}

\item{log}{logical; if TRUE, log density is returned}
}
\value{
density, probability, quantile value or random variable from Rayleigh distribution
}
\description{
Rayleigh distribution
}
\examples{
qrayleigh(0.5, 0.5)
}
