% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{rMHDR}
\alias{rMHDR}
\title{Multivariate version of the Hubbard Decision Research (HDR) Pseudo-Random Number Generator (PRNG)}
\usage{
rMHDR(n, var_ids = NULL, seed = 1)
}
\arguments{
\item{n}{number of samples to return.}

\item{var_ids}{variable IDs}

\item{seed}{seed. Used as a starting index for series of observations}
}
\value{
matrix with columns corresponding to variables
}
\description{
Multivariate version of the Hubbard Decision Research (HDR) Pseudo-Random Number Generator (PRNG)
}
