% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkewNorm.R
\name{rSkewNorm}
\alias{rSkewNorm}
\title{Generate random numbers from Generalized Normal distribution with skew parameter.}
\usage{
rSkewNorm(n, q1, q3, s = 0, lower = -Inf, upper = Inf)
}
\arguments{
\item{n}{number of observations. If `length(n)>1``, the length is taken to be the number required.}

\item{q1}{minimum value}

\item{q3}{maximum value}

\item{s}{skew parameter betweem -1 and 1. When skew parameter equal 0, distribution is normal.}

\item{lower}{lower bound for distribution}

\item{upper}{upper bound for distribution}
}
\value{
a length \code{n} vector of random values.
}
\description{
Generate random numbers from Generalized Normal distribution with skew parameter.
}
\examples{
rSkewNorm(1, 10, 20, 0.5)
}
