% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sech.R, R/utils.R
\name{psech}
\alias{psech}
\alias{dsech}
\alias{qsech}
\alias{fsech}
\alias{dqsech}
\alias{rsech}
\alias{sech}
\title{Hyperbolic secant distribution}
\usage{
psech(x, location = 0, scale = 1)

dsech(x, location = 0, scale = 1)

qsech(p, location = 0, scale = 1)

fsech(p, location = 0, scale = 1)

dqsech(p, location = 0, scale = 1, log = FALSE)

rsech(n, location = 0, scale = 1)

sech(x)
}
\arguments{
\item{x}{numeric vector}

\item{p}{numeric vector of cumulative probabilities}

\item{log}{logical; if TRUE, log density is returned. Default is FALSE}

\item{n}{number of samples to draw}

\item{alpha}{parameter of generalized exponential distribution}

\item{lambda}{parameter of generalized exponential distribution}
}
\value{
cum

vector of hyperbolic secants
}
\description{
Hyperbolic secant distribution with CDF \eqn{F(x)=\frac{2}{\pi}\arctan(\exp(\frac{\pi}{2}\frac{(x-\mu)}{\sigma}))} where \eqn{\mu} and \eqn{\sigma} are location and scale, respectively.
}
\examples{
pgenexp(1, 1, 1)
}
\keyword{internal}
