% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myerson.R
\name{sechMyerson}
\alias{sechMyerson}
\alias{qsechMyerson}
\alias{fsechMyerson}
\alias{dqsechMyerson}
\alias{rsechMyerson}
\alias{psechMyerson}
\alias{dsechMyerson}
\title{Density function for Hyperbolic Secant Myerson distribution (Myerson distribution with hyperbolic secant base function).}
\usage{
qsechMyerson(p, q1, q2, q3, alpha = 0.25)

fsechMyerson(p, q1, q2, q3, alpha = 0.25, log = FALSE)

dqsechMyerson(p, q1, q2, q3, alpha = 0.25, log = FALSE)

rsechMyerson(n, q1, q2, q3, alpha = 0.25)

psechMyerson(q, q1, q2, q3, alpha = 0.25)

dsechMyerson(x, q1, q2, q3, alpha = 0.25, log = FALSE)
}
\arguments{
\item{p}{vector of probabilities}

\item{q1, q2, q3}{numeric values representing bottom quantile, middle quantile (50th percentile)
and top quantile. Quantiles are assumed to be symmetrical.}

\item{alpha}{numerical fixed proportion of distribution below the bottom quantile.
Default value is 0.25}

\item{log}{logical; if TRUE, log density is returned}

\item{n}{number of observations. If `length(n)>1``, the length is taken to be the number required.}

\item{x, q}{vector of quantiles.}
}
\value{
a vector of probabilities of length equal to \code{length(x)}

a vector of exceedance probabilities of length equal to \code{length(x)}.
}
\description{
Density function for Hyperbolic Secant Myerson distribution (Myerson distribution with hyperbolic secant base function).
}
\examples{
qsechMyerson(0.25, 10, 20, 40)
fsechMyerson(0.25, 10, 20, 40)
dqsechMyerson(0.25, 10, 20, 40)
rsechMyerson(1, 10, 20, 40)
psechMyerson(0.25, 10, 20, 40)
dsechMyerson(15, 10, 20, 40)
}
