% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{seq_grid}
\alias{seq_grid}
\title{internal function for making a scaled equispaced, non-bounded grid}
\usage{
seq_grid(f, t, n)
}
\arguments{
\item{f}{from value}

\item{t}{to value}

\item{n}{number of equispaced values}
}
\description{
internal function for making a scaled equispaced, non-bounded grid
}
\keyword{internal}
