% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logis.R
\name{qslogis}
\alias{qslogis}
\alias{fslogis}
\alias{dqslogis}
\alias{rslogis}
\title{Quantile function for Gilchrist's skew-logistic distribution}
\usage{
qslogis(p, location = 0, scale = 1, dlt = 0.5)

fslogis(p, location = 0, scale = 1, dlt = 0.5, log = FALSE)

dqslogis(p, location = 0, scale = 1, dlt = 0.5, log = FALSE)

rslogis(n, location, scale, dlt)
}
\arguments{
\item{p}{numeric probability}

\item{location}{numeric parameter of logistic distribution. Default is 0}

\item{scale}{numeric parameter of logistic distribution. Default is 1}

\item{dlt}{parameter delta of SLD distribution(mixing parameter). Should be within the interval (0,1), default is 0.5}

\item{log}{logical, should the result be returned as a log. Default is FALSE}

\item{n}{number of random variates to generate}
}
\value{
returns QF (\code{qslogis}), QDF (\code{fslogis}), DQF (\code{dqslogis}) or random variate of skew-logistic distribution
}
\description{
Calculate QF, QDF, DQF \code{qslogis}, \code{flogis}, \code{dqlogis} for the skew-logistic distribution
}
\examples{
p_grd <- make_pgrid()
fslogis(p_grd)
dqslogis(p_grd)
}
