% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slogis-xpower.R
\name{fit_slogis_xpower}
\alias{fit_slogis_xpower}
\alias{approx_slogis_xpower}
\alias{qslogis_xpower}
\alias{fslogis_xpower}
\alias{dqslogis_xpower}
\alias{rslogis_xpower}
\title{Fitting the SLxP functions}
\usage{
fit_slogis_xpower(p, q, log.p = FALSE, tol = .Machine$double.eps^2)

approx_slogis_xpower(
  q,
  nterms = 3L,
  p = NULL,
  thin = FALSE,
  n_grid = 1000,
  s_grid = 2L,
  tol = .Machine$double.eps^2
)

qslogis_xpower(u, a)

fslogis_xpower(u, a, log = FALSE)

dqslogis_xpower(u, a, log = FALSE)

rslogis_xpower(n, a)
}
\arguments{
\item{p}{the grid of probability values the quantiles q correspond to. This would be specified if metalog is fitted to the empirical CDF. Default is NULL.}

\item{q}{vector of quantile values (data)}

\item{log.p}{are probabilities provided on log scale. Default FALSE}

\item{tol}{tolerance for solve() and qr.solve(), default is .Machine$double.eps^2}

\item{nterms}{integer number of terms for approximating metalog. Default is 3}

\item{thin}{logical. Should original data be thinned. Default is FALSE.}

\item{n_grid}{in case data thinning is performed, integer number of quantiles to extract from data, if data vector \code{q} is longer than this value}

\item{s_grid}{in case data thinning is performed, probability grid shape parameter passed to \code{qpd::make_pgrid()}. Default is 10.}

\item{u}{numeric probability}

\item{a}{numeric vector of SLxP-QPD parameters.}

\item{log}{logical, should the result be returned as a log. Default is FALSE}

\item{n}{number of samples to draw}
}
\value{
returns QF (\code{qslogis_xpower}), QDF (\code{fslogis_xpower}), DQF (\code{dqslogis_xpower}) or random variate of skew-logistic distribution
}
\description{
Functions for fitting and sampling from SLxP distribution

Calculate QF, QDF, DQF \code{qslogis_xpower}, \code{fslogis_xpower}, \code{dqslogis_xpower} for the SLxP-QPD
}
\details{
\code{fit_slogis_xpower} is for fitting the SLxP function to the set of QP values.
Number of SLxP terms will match the number of QP pairs.
\code{approx_slogis_xpower} is for approximating SLxP function to the set of data.
\code{is_slogis_xpower_valid} is a function for checking if the SLxP is valid
}
\examples{
p <- c(0.1, 0.5, 0.9)
q <- c(4, 9, 12)
a <-fit_slogis_xpower(p,q)
p_grd <- make_pgrid()

fslogis_xpower(p_grd, a)
dqslogis_xpower(p_grd, a)
}
