% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_int}
\alias{split_int}
\title{internal function for splitting the integer value into a homogenous vector}
\usage{
split_int(x, d)
}
\arguments{
\item{x}{integer value to split}

\item{d}{integer divisor}
}
\value{
vector of length d of integers
}
\description{
internal function for splitting the integer value into a homogenous vector
}
\keyword{internal}
