% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqn.R
\name{qsqn}
\alias{qsqn}
\alias{fsqn}
\alias{dqsqn}
\alias{psqn}
\alias{rsqn}
\title{sqn distribution functions}
\usage{
qsqn(p, a)

fsqn(p, a, log = FALSE)

dqsqn(p, a, log = FALSE)

psqn(q, a, n_grid = 50L, s_grid = 2L, tol = 1e-15, maxiter = 1000)

rsqn(n, a)
}
\arguments{
\item{p}{vector of cumulative probabilities corresponding to quantile values}

\item{a}{vector of \code{a}-coefficient parameters of sqn distribution}

\item{log}{should the log density be returned. Default FALSE}

\item{q}{real vector of values}

\item{n_grid}{integer size of helper grid to be passed to \code{make_pgrid}. Default is 50}

\item{s_grid}{integer beta shape of helper grid to be passed to \code{make_pgrid}. Default is 2}

\item{tol}{tolerance value, default is 1e-6}

\item{maxiter}{maximum number of iterations for approximation, default is 1e6}

\item{n}{integer value correponding to number of samples to draw}
}
\description{
Functions for sampling from sqn distribution
}
\details{
\code{qsqn} is a quantile function.
\code{fsqn} is a quantile density function q(u).
\code{dqsqn} is the reciprocal of it is density quantile function f(Q(p)).
\code{psqn} is an approximation of the cumulative density function.
\code{rsqn} is an RNG.
}
\examples{
a <- c(9,  1.8, -1.13, 9)
p <- c(0.1, 0.5, 0.9)
qsqn(p, a)
fsqn(p, a)
dqsqn(p, a)
x <- c(5, 9, 14)
psqn(x, a)
a <- c(2.4, 0.4, -0.08, 9)
rsqn(100, a)
}
