% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlambda.R
\name{tlambda}
\alias{tlambda}
\alias{qtlambda}
\alias{ftlambda}
\alias{dqtlambda}
\alias{ptlambda}
\alias{rtlambda}
\title{Tukey lambda distribption}
\usage{
qtlambda(p, lambda)

ftlambda(p, lambda)

dqtlambda(p, lambda, log = FALSE)

ptlambda(q, ..., lower = 0, upper = 1, tol = 1e-06, silent = TRUE, trace = 0)

rtlambda(n, lambda)
}
\arguments{
\item{p}{vector of probabilities}

\item{lambda}{parameter beta of Tukey lambda distribution (scale and shape).}

\item{log}{logical; if TRUE, log density is returnes. Default is FALSE}

\item{q}{vector of quantiles}

\item{...}{used by method}

\item{lower, upper}{the \code{stats::uniroot} lower and upper end points of the interval to be searched. Defaults are 0 and 1, respectively}

\item{tol}{the \code{stats::uniroot} desired accuracy (convergence tolerance). Default value 1e-06}

\item{silent}{the \code{base::try} argument. Default is TRUE}

\item{trace}{integer number passed to \code{stats::uniroot}; if positive, tracing information is produced. Higher values giving more details.}

\item{n}{numeric; number of samples to draw from Tukey Lambda distribution}
}
\value{
vector
}
\description{
Defines quantile function (ICDF), and quantile density function as well as probability function (CDF) for Tukey lambda distribution
}
\examples{
qtlambda(0.1, -0.5)
p <- runif(1e4)
x <- qtlambda(p, -0.5)+qtlambda(p, 0.5)
hist(x,30)
}
