% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{make_ecdf_df}
\alias{make_ecdf_df}
\title{Compute ECDF for a sample}
\usage{
make_ecdf_df(x)
}
\arguments{
\item{x}{numeric sample}
}
\value{
data frame with 2 columns p - vector of empirical probabilities, q - vector of sorted original sample values
}
\description{
Function for creating an empirical CDF from a numerical sample
}
\examples{
make_ecdf_df(runif(10))

}
