% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalog.R
\name{fit_metalog}
\alias{fit_metalog}
\alias{approx_metalog}
\alias{approx_max_metalog}
\alias{is_metalog_valid}
\title{Fitting the metalog functions}
\usage{
fit_metalog(p, q, bl = NULL, bu = NULL, log.p = FALSE)

approx_metalog(
  q,
  nterms = 3L,
  bl = NULL,
  bu = NULL,
  thin = FALSE,
  n_grid = 1000,
  s_grid = 2L,
  tol = 1e-07
)

approx_max_metalog(
  q,
  bl = NULL,
  bu = NULL,
  thin = FALSE,
  n_grid = 1000,
  s_grid = 2L,
  tol = 1e-07
)

is_metalog_valid(a, bl = NULL, bu = NULL)
}
\arguments{
\item{p}{vector of cumulative probabilities corresponding to quantile values}

\item{q}{vector of quantile values (data)}

\item{bl}{real value of lower boundary (for bounded metalog). Default NULL}

\item{bu}{real value of upper boundary (for bounded metalog). Default NULL}

\item{log.p}{are probabilities provided on log scale. Default FALSE}

\item{nterms}{integer number of terms for approximating metalog. Default is 3}

\item{thin}{logical. Should original data be thinned. Default is FALSE.}

\item{n_grid}{in case data thinning is performed, integer number of quantiles to extract from data, if data vector \code{q} is longer than this value}

\item{s_grid}{in case data thinning is performed, probability grid shape parameter passed to \code{qpd::make_pgrid()}. Default is 10.}

\item{tol}{tolerance for qr.solve()}

\item{a}{vector of \code{a}-coefficient parameters of metalog distribution}
}
\description{
Functions for fitting and sampling from metalog distribution
}
\details{
\code{fit_metalog} is for fitting the metalog function to the set of QP values.
Number of metalog terms will match the number of QP pairs.
\code{approx_metalog} is for approximating metalog function to the set of data.
\code{is_metalog_valid} is a function for checking if the metalog is valid
}
\examples{
p <- c(0.1, 0.5, 0.9)
q <- c(4, 9, 12)
fit_metalog(p,q)
a <- c(9,  1.8, -1.13)
is_metalog_valid(a)
}
