% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfld.R
\name{fld}
\alias{fld}
\alias{qgfld}
\alias{fgfld}
\alias{dqgfld}
\alias{rgfld}
\alias{pgfld}
\title{Generalized Flattened Logistic Distribution (FLD)}
\usage{
qgfld(p, bt, k, dlt = 0.5, a = 0)

fgfld(p, bt, k, dlt = 0.5, a = 0)

dqgfld(p, bt, k, dlt = 0.5, a = 0, log = FALSE)

rgfld(n, bt, k, dlt = 0.5, a = 0)

pgfld(q, bt, k, dlt = 0.5, a = a, tol = 1e-06)
}
\arguments{
\item{p}{vector of probabilities}

\item{bt}{parameter beta of GFLD distribution (scale). Should be non-negative.}

\item{k}{parameter k of GFLD distribution (shape). Should be non-negative.}

\item{dlt}{parameter delta of GFLD distribution(mixing parameter). Should be within the interval (0,1), default is 0.5}

\item{a}{location parameter alpha of GFLD distribution(location parameter), default is 0}

\item{log}{logical; if TRUE, log density is returnes. Default is FALSE}

\item{n}{numeric; number of samples to draw from FLD distribution}

\item{q}{vector of quantiles}

\item{tol}{tolerance value for optimization. Default value 1e-06}
}
\value{
vector
}
\description{
Defines quantile function (ICDF), probability density and quantile density function as well as probability function (CDF) for generalized flattened logistic distribution
}
\examples{
qgfld(0.1, 0.5, 0.3, 0.5)
# centered gfld
p <- runif(1e4)
x <- qgfld(p, 0.25, 1)-qgfld(0.5, 0.25, 1)
hist(x,30)
}
