% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsld.R
\name{fld}
\alias{fld}
\alias{qfld}
\alias{ffld}
\alias{dqfld}
\alias{rfld}
\alias{pfld}
\title{Flattened Logistic Distribution (FLD)}
\usage{
qfld(p, bt, k, a = 0)

ffld(p, bt, k, a = 0)

dqfld(p, bt, k, a = 0, log = FALSE)

rfld(n, bt, k, a = 0)

pfld(q, ..., lower = 0, upper = 1, tol = 1e-06, silent = TRUE, trace = 0)
}
\arguments{
\item{p}{vector of probabilities}

\item{bt}{parameter beta of GFLD distribution (scale). Should be non-negative.}

\item{k}{parameter k of GFLD distribution (shape). Should be non-negative.}

\item{a}{location parameter alpha of GFLD distribution(location parameter), default is 0}

\item{log}{logical; if TRUE, log density is returnes. Default is FALSE}

\item{n}{numeric; number of samples to draw from FLD distribution}

\item{q}{vector of quantiles}

\item{...}{used by method}

\item{lower, upper}{the \code{stats::uniroot} lower and upper end points of the interval to be searched. Defaults are 0 and 1, respectively}

\item{tol}{the \code{stats::uniroot} desired accuracy (convergence tolerance). Default value 1e-06}

\item{silent}{the \code{base::try} argument. Default is TRUE}

\item{trace}{integer number passed to \code{stats::uniroot}; if positive, tracing information is produced. Higher values giving more details.}
}
\value{
vector
}
\description{
Defines quantile function (ICDF), probability density and quantile density function as well as probability function (CDF) for flattened logistic distribution
}
\examples{
qfld(0.1, 0.5, 0.3, 0.5)
# centered gfld
p <- runif(1e4)
x <- qfld(p, 0.25, 1)-qfld(0.5, 0.25, 1)
hist(x,30)
}
