% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gldfkml.R
\name{qGLDfkml}
\alias{qGLDfkml}
\alias{fGLDfkml}
\alias{dqGLDfkml}
\alias{rGLDfkml}
\alias{pGLDfkml}
\alias{is_GLDfkml_valid}
\alias{aGLDfkml_mean}
\alias{aGLDfkml_median}
\title{Generalized Lambda Distribution (GLD) FKML parameterization}
\usage{
qGLDfkml(u, l1, l2, l3, l4)

fGLDfkml(u, l1, l2, l3, l4)

dqGLDfkml(u, l1, l2, l3, l4, log = FALSE)

rGLDfkml(n, l1, l2, l3, l4)

pGLDfkml(q, ..., lower = 0, upper = 1, tol = 1e-06, silent = TRUE, trace = 0)

is_GLDfkml_valid(l1, l2, l3, l4)

aGLDfkml_mean(l1, l2, l3, l4)

aGLDfkml_median(l1, l2, l3, l4)
}
\arguments{
\item{u}{numeric vector of probabilities}

\item{l1}{GLD parameter \eqn{\lambda_1}, (FKML parameterization)}

\item{l2}{GLD parameter \eqn{\lambda_2}, (FKML parameterization)}

\item{l3}{GLD parameter \eqn{\lambda_3}, (FKML parameterization)}

\item{l4}{GLD parameter \eqn{\lambda_4}, (FKML parameterization)}

\item{log}{should the log density be returned. Default=FALSE}

\item{n}{numeric number of samples to draw}

\item{q}{vector of quantiles}

\item{...}{used by method}

\item{lower, upper}{the \code{stats::uniroot} lower and upper end points of the interval to be searched. Defaults are 0 and 1, respectively}

\item{tol}{the \code{stats::uniroot} desired accuracy (convergence tolerance). Default value 1e-06}

\item{silent}{the \code{base::try} argument. Default is TRUE}

\item{trace}{integer number passed to \code{stats::uniroot}; if positive, tracing information is produced. Higher values giving more details.}
}
\value{
quantiles, QDF, DQF, random samples or probabilities of GLD (FKML parameterization)
}
\description{
Quantile function, quantile density, density quantile and inverse quantile
functions for GLD distribution with FKML parameterization.
\code{aGLDfkml_mean} and \code{aGLDfkml_median} are theoretical mean and median, which can be used
for adjusting the quantile likelihood.
}
\examples{
p_grd <- make_pgrid()
qGLDfkml(p_grd, 1, 1, -1/8, -1/32)
}
