% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gldrs.R
\name{qGLDrs}
\alias{qGLDrs}
\alias{fGLDrs}
\alias{dqGLDrs}
\alias{rGLDrs}
\alias{pGLDrs}
\alias{is_GLDrs_valid}
\title{Generalized Lambda Distribution (GLD) RS parameterization}
\usage{
qGLDrs(u, l1, l2, l3, l4)

fGLDrs(u, l1, l2, l3, l4, log = FALSE)

dqGLDrs(u, l1, l2, l3, l4, log = FALSE)

rGLDrs(n, l1, l2, l3, l4)

pGLDrs(
  q,
  l1,
  l2,
  l3,
  l4,
  n_grid = 50L,
  s_grid = 5L,
  tol = 1e-15,
  maxiter = 1000
)

is_GLDrs_valid(l1, l2, l3, l4)
}
\arguments{
\item{u}{numeric vector of probabilities}

\item{l1}{GLD parameter \eqn{\lambda_1}}

\item{l2}{GLD parameter \eqn{\lambda_2}}

\item{l3}{GLD parameter \eqn{\lambda_3}}

\item{l4}{GLD parameter \eqn{\lambda_4}}

\item{log}{should the log density be returned. Default=FALSE}

\item{n}{numeric number of samples to draw}

\item{q}{quantile value for which the corresponding cumulative probability value should be found.}

\item{n_grid}{integer size of helper grid to be passed to \code{make_pgrid}. Default is 50}

\item{s_grid}{integer beta shape of helper grid to be passed to \code{make_pgrid}. Default is 5}

\item{tol}{numeric tolerance balue for approximating CDF. Default 1e-15}

\item{maxiter}{numeric maximum number of iteration}
}
\value{
quantiles, QDF, DQF, random samples or probabilities of GLD
}
\description{
Quantile function, quantile density, density quantile and inverse quantile
functions for GLD distribution with RS parameterization.
\code{aGLDrs_mean} and \code{aGLDrs_median} are theoretical mean and median, which can be used
for adjusting the quantile likelihood.
}
\examples{
p_grd <- make_pgrid()
is_GLDrs_valid(1, -1, -1/8, -1/32)
qGLDrs(p_grd, 1, -1, -1/8, -1/32)
}
