% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnk.R
\name{qgnk}
\alias{qgnk}
\alias{fgnk}
\alias{dqgnk}
\alias{rgnk}
\alias{pgnk}
\alias{is_gnk_valid}
\alias{is_gnk_valid_u}
\title{Functions for g-and-k distribution}
\usage{
qgnk(p, A, B, C = 0.8, g, k, zscale = FALSE)

fgnk(p, A, B, C = 0.8, g, k, log = FALSE, zscale = FALSE)

dqgnk(p, A, B, C = 0.8, g, k, log = FALSE, zscale = FALSE)

rgnk(n, A, B, C = 0.8, g, k)

pgnk(
  q,
  A,
  B,
  C = 0.8,
  g,
  k,
  n_grid = 100L,
  s_grid = 5L,
  tol = 1e-15,
  maxiter = 1000,
  zscale = FALSE
)

is_gnk_valid(A, B, C = 0.8, g, k)

is_gnk_valid_u(A, B, C = 0.8, g, k, maxiter = 100L)
}
\arguments{
\item{p}{numeric vector of probabilities}

\item{A}{numeric location parameter of g-and-k distribution}

\item{B}{positive numeric scale parameter of g-and-k distribution}

\item{C}{numeric parameter of g-and-k distribution. Default is 0.8}

\item{g}{numeric skeweness parameter of g-and-k distribution}

\item{k}{non-negative numeric kurtosis parameter of g-and-k distribution}

\item{zscale}{is the probability input provided on scale of \code{z} values, i.e. as \code{z=qnorm(p)}. Default is FALSE}

\item{log}{logical should the result be returned as log(). Default is FALSE}

\item{n}{numeric number of samples to draw}

\item{q}{quantile value for which the corresponding cumulative probability value should be found.}

\item{n_grid}{integer size of helper grid to be passed to \code{make_pgrid}. Default is 50}

\item{s_grid}{integer beta shape of helper grid to be passed to \code{make_pgrid}. Default is 5}

\item{tol}{numeric tolerance balue for approximating CDF. Default 1e-15}

\item{maxiter}{numeric maximum number of iteration}
}
\value{
vector of quantiles for g-and-k distribution

vector of quantiles for g-and-h distribution
}
\description{
Functions for g-and-k distribution
}
\examples{
qgnk(0.1, 3,1,0.8, 2,0.5)
}
