% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govindarajulu.R
\name{govindarajulu}
\alias{govindarajulu}
\alias{qgovindarajulu}
\alias{fgovindarajulu}
\alias{dqgovindarajulu}
\alias{pgovindarajulu}
\alias{dgovindarajulu}
\title{Govindarajulu probability distribution}
\usage{
qgovindarajulu(p, sg, bt)

fgovindarajulu(p, sg, bt)

dqgovindarajulu(p, sg, bt, log = FALSE)

pgovindarajulu(
  q,
  ...,
  lower = 0,
  upper = 1,
  tol = 1e-06,
  silent = TRUE,
  trace = 0
)

dgovindarajulu(x, sg, bt, tol = 1e-06)
}
\arguments{
\item{p}{vector of probabilities}

\item{sg}{parameter sigma of Govindarajulu distribution}

\item{bt}{parameter beta of Govindarajulu distribution}

\item{log}{logical; if TRUE, log density is returned. Default is FALSE}

\item{q}{vector of quantiles}

\item{...}{used by method}

\item{lower, upper}{the \code{stats::uniroot} lower and upper end points of the interval to be searched. Defaults are 0 and 1, respectively}

\item{tol}{the \code{stats::uniroot} desired accuracy (convergence tolerance). Default value 1e-06}

\item{silent}{the \code{base::try} argument. Default is TRUE}

\item{trace}{integer number passed to \code{stats::uniroot}; if positive, tracing information is produced. Higher values giving more details.}

\item{x}{numeric vector}
}
\value{
vector
}
\description{
Defines quantile function (ICDF), probability density and quantile density function as well as probability function (CDF) for Govindarajulu distribution
}
\examples{
qgovindarajulu(0.1, 1, 5)
fgovindarajulu((1:5)/10, 1, 5)
}
