% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{qhalfnorm}
\alias{qhalfnorm}
\alias{phalfnorm}
\alias{dhalfnorm}
\title{Half normal distribution}
\usage{
qhalfnorm(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

phalfnorm(x, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

dhalfnorm(x, location = 0, scale = 1, log = FALSE)
}
\arguments{
\item{p}{numeric probability}

\item{location}{numeric parameter of halfnormal distribution. Default is 0}

\item{scale}{numeric parameter of halfnormal distribution. Default is 1}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X\leq x]}, otherwise \eqn{P[X>x]}}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p). Default is FALSE}

\item{x}{numeric value}

\item{log}{logical, should the result be returned as a log. Default is FALSE}
}
\value{
returns QDF (\code{fnorm}) and DQF (\code{dqnorm}) of normal distribution
}
\description{
Half normal distribution
}
\examples{
p_grd <- make_pgrid()
qhalfnorm(p_grd)
}
