% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqf.R
\name{iqf}
\alias{iqf}
\title{Function factory for creating an inverse quantile function}
\usage{
iqf(QFUN)
}
\arguments{
\item{QFUN}{bare name of the quantile function to be inverted. Assumes that the first argument in the quantile function is \code{p} for probability.}
}
\value{
function. Approximated CDF for a given quantile function.
The inverse function will have the following additional arguments, passed to stats::uniroot()
lower=0, upper=1, tol=1e-6, silent=TRUE.
The resulting inverse quantile function is fully vectorized with regards to
all of its arguments (shorter vectors are recycled).
}
\description{
Function factory for creating an inverse quantile function
}
\examples{
qmyfun <- function(p, b, k){b*(log(p/(1-p))+k*p)}
pmyfun <- iqf(qmyfun)
x <- qmyfun(1:9/10, 2, 3)
pmyfun(x, 2, 3)
}
