% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalog.R
\name{qmetalog}
\alias{qmetalog}
\alias{fmetalog}
\alias{pmetalog}
\alias{rmetalog}
\title{Metalog distribution functions}
\usage{
qmetalog(p, a, bl = NULL, bu = NULL)

fmetalog(p, a, bl = NULL, bu = NULL)

pmetalog(x, a, bl = NULL, bu = NULL, tol = 1e-06, maxiter = 1e+06)

rmetalog(n, a, bl = NULL, bu = NULL)
}
\arguments{
\item{p}{vector of cumulative probabilities corresponding to quantile values}

\item{a}{vector of \code{a}-coefficient parameters of metalog distribution}

\item{bl}{real value of lower boundary (for bounded metalog). Default NULL}

\item{bu}{real value of upper boundary (for bounded metalog). Default NULL}

\item{x}{real vector of values}

\item{tol}{tolerance value, default is 1e-6}

\item{maxiter}{maximum number of iterations for approximation, default is 1e6}

\item{n}{integer value correponding to number of samples to draw}
}
\description{
Functions for sampling from metalog distribution
}
\details{
\code{qmetalog} is a quantile function.
\code{fmetalog} is a quantile density function q(u). The reciprocal of it is density quantile function f(Q(p)).
\code{pmetalog} is an approximation of the cumulative density function.
\code{rmetalog} is an RNG.
}
\examples{
a <- c(9,  1.8, -1.13)
p <- c(0.1, 0.5, 0.9)
qmetalog(p, a)
fmetalog(p, a)
x <- c(5, 9, 14)
pmetalog(x, a)
a <- c(2.4, 0.4, -0.08)
rmetalog(100, a)
}
