% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metalog.R
\name{qmetalog}
\alias{qmetalog}
\alias{fmetalog}
\alias{dqmetalog}
\alias{pmetalog}
\alias{rmetalog}
\title{Metalog distribution functions}
\usage{
qmetalog(p, a, bl = -Inf, bu = Inf, log.p = FALSE)

fmetalog(p, a, bl = -Inf, bu = Inf, log.p = FALSE, log = FALSE)

dqmetalog(p, a, bl = -Inf, bu = Inf, log.p = FALSE, log = FALSE)

pmetalog(
  q,
  a,
  bl = -Inf,
  bu = Inf,
  n_grid = 50L,
  s_grid = 2L,
  tol = 1e-15,
  maxiter = 1000,
  log.p = FALSE
)

rmetalog(n, a, bl = -Inf, bu = Inf)
}
\arguments{
\item{p}{vector of cumulative probabilities corresponding to quantile values}

\item{a}{vector of \code{a}-coefficient parameters of metalog distribution}

\item{bl}{real value of lower boundary (for bounded metalog). Default -Inf}

\item{bu}{real value of upper boundary (for bounded metalog). Default Inf}

\item{log.p}{should log probability be returned}

\item{log}{should the log density be returned. Default FALSE}

\item{q}{real vector of values}

\item{n_grid}{integer size of helper grid to be passed to \code{make_pgrid}. Default is 50}

\item{s_grid}{integer beta shape of helper grid to be passed to \code{make_pgrid}. Default is 2}

\item{tol}{tolerance value, default is 1e-6}

\item{maxiter}{maximum number of iterations for approximation, default is 1e6}

\item{n}{integer value correponding to number of samples to draw}
}
\description{
Functions for sampling from metalog distribution
}
\details{
\code{qmetalog} is a quantile function.
\code{fmetalog} is a quantile density function q(u).
\code{dqmetalog} is the reciprocal of it is density quantile function f(Q(p)).
\code{pmetalog} is an approximation of the cumulative density function.
\code{rmetalog} is an RNG.
}
\examples{
a <- c(9,  1.8, -1.13)
p <- c(0.1, 0.5, 0.9)
qmetalog(p, a)
fmetalog(p, a)
dqmetalog(p, a)
x <- c(5, 9, 14)
pmetalog(x, a)
a <- c(2.4, 0.4, -0.08)
rmetalog(100, a)
}
