% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpd-package.R
\docType{package}
\name{qpd}
\alias{qpd}
\title{qpd: A package for quantile parametrized distributions in R}
\description{
The qpd package provides functions for samping quantile parametrized distibutions.
}
\section{qpd functions}{

The qpd functions implement Johnson QPD, Generalized Johnson QPD, Myerson distribution, Skew-Normal QPD, Q-normal. For metalog distribution, please consult rmetalog package.
}

