% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayleigh.R
\name{drayleigh2}
\alias{drayleigh2}
\alias{prayleigh2}
\alias{qrayleigh2}
\alias{rrayleigh2}
\alias{frayleigh2}
\alias{dqrayleigh2}
\alias{ffrayleigh2}
\title{Two-parameter Rayleigh distribution}
\usage{
drayleigh2(x, mu, lambda)

prayleigh2(x, mu, lambda)

qrayleigh2(p, mu, lambda)

rrayleigh2(n, mu, lambda)

frayleigh2(p, mu, lambda)

dqrayleigh2(p, mu, lambda)

ffrayleigh2(p, mu, lambda)
}
\arguments{
\item{x}{numeric vector of data}

\item{mu}{numeric location parameter of two-parameter Rayleigh distribution}

\item{lambda}{numeric scale parameter of two-parameter Rayleigh distribution}

\item{p}{numeric vector of probabilities}

\item{n}{numeric number of random values to draw}
}
\value{
density, probability, quantile value or random variable from  two-parameter Rayleigh distribution
}
\description{
Two-parameter Rayleigh distribution
}
\examples{
qrayleigh2(0.5, 0.5, 0.5)
}
