% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sech}
\alias{sech}
\title{internal function for hyperbolic secant borrowed from {pracma}}
\usage{
sech(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
vector of hyperbolic secants
}
\description{
internal function for hyperbolic secant borrowed from {pracma}
}
\keyword{internal}
