% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakeby.R
\name{qwakeby}
\alias{qwakeby}
\alias{fwakeby}
\alias{dqwakeby}
\alias{rwakeby}
\alias{pwakeby}
\alias{dwakeby}
\alias{is_wakeby_valid}
\alias{ffwakeby}
\title{Wakeby distribution}
\usage{
qwakeby(u, alpha, beta, gamma, delta, xi)

fwakeby(u, alpha, beta, gamma, delta, xi, log = FALSE)

dqwakeby(u, alpha, beta, gamma, delta, xi, log = FALSE)

rwakeby(n, alpha, beta, gamma, delta, xi)

pwakeby(q, ..., lower = 0, upper = 1, tol = 1e-06, silent = TRUE, trace = 0)

dwakeby(x, alpha, beta, gamma, delta, xi, tol = 1e-06, log = FALSE)

is_wakeby_valid(alpha, beta, gamma, delta, xi, n_grid = 100L, s_grid = 2L)

ffwakeby(u, alpha, beta, gamma, delta, xi)
}
\arguments{
\item{u}{numeric vector of probabilities}

\item{alpha}{Wakeby scale parameter \eqn{\alpha}}

\item{beta}{Wakeby shape parameter \eqn{\beta}}

\item{gamma}{Wakeby scale parameter \eqn{\gamma}}

\item{delta}{Wakeby shape parameter \eqn{\delta}}

\item{xi}{Wakeby location parameter \eqn{\xi}}

\item{log}{should the log density be returned}

\item{n}{numeric number of samples to draw}

\item{q}{vector of quantiles}

\item{...}{used by method}

\item{lower, upper}{the \code{stats::uniroot} lower and upper end points of the interval to be searched. Defaults are 0 and 1, respectively}

\item{tol}{the \code{stats::uniroot} desired accuracy (convergence tolerance). Default value 1e-06}

\item{silent}{the \code{base::try} argument. Default is TRUE}

\item{trace}{integer number passed to \code{stats::uniroot}; if positive, tracing information is produced. Higher values giving more details.}

\item{x}{numeric vector of data}

\item{n_grid}{integer size of helper grid to be passed to \code{make_pgrid}. Default is 50}

\item{s_grid}{integer beta shape of helper grid to be passed to \code{make_pgrid}. Default is 5}
}
\value{
quantiles, QDF, DQF, random samples or probabilities of Wakeby disribution
}
\description{
Wakeby distribution
}
\details{
Support \eqn{\xi} to \eqn{\infty} if  \eqn{\delta \geq 0, \gamma>0} or \eqn{\xi} to \eqn{\xi+(\alpha/\beta)-(\gamma/\delta)} otherwise.
}
\examples{
p_grd <- make_pgrid()
q_grd <- qwakeby(p_grd, 5, 3, 0.1, 0.2, 0)
f_grd <- fwakeby(p_grd, 5, 3, 0.1, 0.2, 0)
dq_grd <- dqwakeby(p_grd, 5, 3, 0.1, 0.2, 0)
d_grd <- dwakeby(q_grd,5, 3, 0.1, 0.2, 0)
plot(q_grd,dq_grd, type="l", lwd=1, lty=2)
lines(q_grd,d_grd, col="firebrick")
is_wakeby_valid(10,5,1,0.3,0)
}
